--- flutter/wtc/lib/intro_screens/intro_page2.dart
@@ -1,24 +1,0 @@
-import 'package:flutter/material.dart';
-
-class IntroPage2 extends StatefulWidget {
-  const IntroPage2({super.key});
-
-  @override
-  State<IntroPage2> createState() => _IntroPage2State();
-}
-
-class _IntroPage2State extends State<IntroPage2> {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      color: Colors.grey.shade300,
-      alignment: Alignment.center,
-      child: const Text(
-        "Intro Page 2",
-        style: TextStyle(
-          color: Colors.black,
-        ),
-      ),
-    );
-  }
-}
