+++ flutter/wtc/lib/pages/intro_screens/intro_page1.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class IntroPage1 extends StatefulWidget {
+  const IntroPage1({super.key});
+
+  @override
+  State<IntroPage1> createState() => _IntroPage1State();
+}
+
+class _IntroPage1State extends State<IntroPage1> {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: Colors.white,
+      alignment: Alignment.center,
+      child: const Text("Intro Page 1"),
+    );
+  }
+}
